converted class component to functional component and Added separate file render.js and settings.js
