+++ assets/css/blocks/advanced-heading.css
@@ -1,1 +1,1 @@
-.wp-block-uagb-advanced-heading{padding:0;margin:0 auto}.wp-block-uagb-advanced-heading .uagb-heading-text{margin:0;text-align:center}.wp-block-uagb-advanced-heading .uagb-separator-wrap{font-size:0;text-align:center}.wp-block-uagb-advanced-heading .uagb-separator{border-top-style:solid;display:inline-block;border-top-width:2px;width:5%;margin:0px 0px 10px 0px}.wp-block-uagb-advanced-heading .uagb-desc-text{margin:0;text-align:center}
+.wp-block-uagb-advanced-heading{padding:0;margin:0 auto}.wp-block-uagb-advanced-heading .uagb-heading-text .uagb-desc-text{margin:0}.wp-block-uagb-advanced-heading .uagb-separator-wrap{font-size:0}.wp-block-uagb-advanced-heading .uagb-separator{border-top-style:solid;display:inline-block;border-top-width:2px;width:5%;margin:0px 0px 10px 0px}
