+++ src/blocks/advanced-heading/style.scss
@@ -7,27 +7,17 @@
 .wp-block-uagb-advanced-heading  {
 	padding: 0;
 	margin: 0 auto;
-
-	.uagb-heading-text {
-		margin: 0;
-		text-align : center;
+	.uagb-heading-text .uagb-desc-text {
+		margin: 0;	
 	}
-
 	.uagb-separator-wrap {
 		font-size: 0;
-		text-align : center;
 	}
-
 	.uagb-separator {
 	    border-top-style: solid;
 	    display: inline-block;
 	    border-top-width: 2px;
 	    width: 5%;
 	    margin: 0px 0px 10px 0px;
-	}
-
-	.uagb-desc-text {
-		margin: 0;
-		text-align : center;
 	}
 }
